@@ -8,35 +8,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.1.3",
-    "@angular/cdk": "^8.1.1",
-    "@angular/common": "~8.1.3",
-    "@angular/compiler": "~8.1.3",
-    "@angular/core": "~8.1.3",
-    "@angular/fire": "^5.2.1",
-    "@angular/forms": "~8.1.3",
-    "@angular/material": "^8.1.1",
-    "@angular/platform-browser": "~8.1.3",
-    "@angular/platform-browser-dynamic": "~8.1.3",
-    "@angular/pwa": "^0.801.2",
-    "@angular/router": "~8.1.3",
-    "@angular/service-worker": "^8.1.3",
-    "@ngxs/storage-plugin": "^3.5.0",
-    "@ngxs/store": "^3.5.0",
-    "firebase": "^6.3.3",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/fire": "^5.2.3",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/pwa": "^0.803.20",
+    "@angular/router": "^8.2.14",
+    "@angular/service-worker": "^8.2.14",
+    "@ngxs/storage-plugin": "^3.6.0",
+    "@ngxs/store": "^3.6.0",
+    "firebase": "^7.6.0",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.3",
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.801.2",
-    "@angular-devkit/build-angular": "^0.801.2",
-    "@angular/cli": "^8.1.2",
-    "@angular/compiler-cli": "~8.1.3",
-    "@angular/language-service": "~8.1.3",
-    "@types/node": "^12.6.8",
-    "codelyzer": "^5.1.0",
-    "tslint": "^5.18.0",
-    "typescript": "^3.4.5"
+    "@angular-devkit/architect": "^0.803.20",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@types/node": "^12.12.17",
+    "codelyzer": "^5.2.0",
+    "tslint": "^6.0.0-beta1",
+    "typescript": "^3.5.3"
   }
 }
