@@ -1,22 +1,24 @@
 import { Injectable } from '@angular/core';
 import { Observable, of, combineLatest } from 'rxjs';
-import { map, take, switchMap } from 'rxjs/operators';
+import { map, take, switchMap, startWith, distinctUntilChanged } from 'rxjs/operators';
 import { Select } from '@ngxs/store';
 
 import { IAuthState } from '@dbase/state/auth.action';
-import { TStateSlice, IAttendState, IPaymentState, IApplicationState, IAdminState } from '@dbase/state/state.define';
+import { TStateSlice, IAttendState, IPaymentState, IApplicationState, IAdminState, IProviderState, IForumState } from '@dbase/state/state.define';
 import { IMemberState, IPlanState, ITimetableState, IState, IAccountState, IUserState } from '@dbase/state/state.define';
-import { joinDoc, sumPayment, sumAttend, calendarDay, buildTimetable, buildPlan, getDefault, getCurrent, getStore, getState } from '@dbase/state/state.library';
+import { joinDoc, sumPayment, sumAttend, calendarDay, buildTimetable, buildPlan, getDefault, getCurrent, getStore, getState, buildProvider } from '@dbase/state/state.library';
 
 import { DBaseModule } from '@dbase/dbase.module';
-import { STORE, FIELD, BONUS } from '@dbase/data/data.define';
+import { STORE, FIELD, BONUS, COLLECTION } from '@dbase/data/data.define';
 import { SORTBY } from '@library/config.define';
-import { IStoreMeta, IRegister, IStatusConnect, IStatusAccount } from '@dbase/data/data.schema';
-import { addWhere } from '@dbase/fire/fire.library';
-import { TWhere } from '@dbase/fire/fire.interface';
+import { IStoreMeta, IRegister, IStatusConnect, IStatusAccount, IForumBase, IComment, IReact } from '@dbase/data/data.schema';
+
+import { FireService } from '@dbase/fire/fire.service';
+import { addWhere, addOrder } from '@dbase/fire/fire.library';
+import { TWhere, IQuery } from '@dbase/fire/fire.interface';
 
 import { asArray } from '@lib/array.library';
-import { DATE_FMT, TDate, getDate } from '@lib/date.library';
+import { Instant, TDate, getDate } from '@lib/instant.library';
 import { cloneObj, sortKeys } from '@lib/object.library';
 import { dbg } from '@lib/logger.library';
 
@@ -26,23 +28,25 @@ import { dbg } from '@lib/logger.library';
  */
 @Injectable({ providedIn: DBaseModule })
 export class StateService {
-	@Select() private auth$!: Observable<IAuthState>;
-	@Select() private client$!: Observable<TStateSlice<IStoreMeta>>;
-	@Select() private member$!: Observable<TStateSlice<IStoreMeta>>;
-	@Select() private attend$!: Observable<TStateSlice<IStoreMeta>>;
-	@Select() private admin$!: Observable<TStateSlice<IStoreMeta>>;
-	@Select() private local$!: Observable<TStateSlice<IStoreMeta>>;
+	@Select() auth$!: Observable<IAuthState>;
+	@Select() client$!: Observable<TStateSlice<IStoreMeta>>;
+	@Select() member$!: Observable<TStateSlice<IStoreMeta>>;
+	@Select() attend$!: Observable<TStateSlice<IStoreMeta>>;
+	@Select() admin$!: Observable<TStateSlice<IStoreMeta>>;
+	@Select() local$!: Observable<TStateSlice<IStoreMeta>>;
+	private forum$: Observable<TStateSlice<IStoreMeta>> = of({ forum: [] });
 
 	private dbg = dbg(this);
 	public states: IState;
 
-	constructor() {
+	constructor(private fire: FireService) {
 		this.states = {                   // a Lookup map for Slice-to-State
 			'client': this.client$,
 			'member': this.member$,
 			'attend': this.attend$,
 			'admin': this.admin$,
 			'local': this.local$,
+			'forum': this.forum$,
 		}
 	}
 
@@ -68,6 +72,7 @@ export class StateService {
 			.then(table => table[0]);				// only the first document
 	}
 
+	/** lookup the Default store for a given 'type' */
 	getDefault(state: IApplicationState, type: string) {
 		return getDefault(state, type);
 	}
@@ -76,12 +81,6 @@ export class StateService {
 		return obs
 			.pipe(take(1))									// only the first snapshot
 			.toPromise();
-	}
-
-	getLatest<T>(...obs: Observable<T>[]) {
-		return combineLatest(...obs)
-			.pipe(take(1))
-			.toPromise()
 	}
 
 	/**
@@ -183,10 +182,29 @@ export class StateService {
 	 * client.price -> has an array of asAt Price documents
 	 */
 	getPlanData(date?: TDate): Observable<IPlanState> {
+		const filterIcon = addWhere(FIELD.type, [STORE.plan, STORE.default]);
+
 		return this.getMemberData(date).pipe(
 			joinDoc(this.states, 'client', STORE.plan, undefined, date),
 			joinDoc(this.states, 'client', STORE.price, undefined, date),
+			joinDoc(this.states, 'client', STORE.icon, filterIcon, date),
 			map(table => buildPlan(table)),
+		)
+	}
+
+	/**
+	 * A list of Sign-in Providers
+	 * client.provider	-> has an array of Provider documents
+	 * client.icon			-> has an array of Provider icons
+	 */
+	getProviderData(date?: TDate): Observable<IProviderState> {
+		const filterProvider = addWhere(FIELD.expire, 0);
+		const filterIcon = addWhere(FIELD.type, [STORE.provider, STORE.default]);
+
+		return this.getMemberData(date).pipe(
+			joinDoc(this.states, 'client', STORE.provider, filterProvider, date),
+			joinDoc(this.states, 'client', STORE.icon, filterIcon, date),
+			map(table => buildProvider(table)),
 		)
 	}
 
@@ -214,6 +232,29 @@ export class StateService {
 	}
 
 	/**
+	 * Observe Forum data direct from FireStore (because it is not persisted to local Store)
+	 * forum.comment	-> has an array of Comments about this date's schedule
+	 * forum.react		-> has an array of Reacts about this date's schedule
+	 */
+	getForumData(date?: TDate): Observable<IForumState> {
+		const query: IQuery = {
+			where: addWhere('track.date', getDate(date).format(Instant.FORMAT.yearMonthDay)),
+			orderBy: addOrder(FIELD.stamp),
+		}
+
+		return this.fire.listen<IForumBase>(COLLECTION.forum, query, 'snapshotChanges').pipe(
+			startWith([] as IForumBase[]),													// dont make subcriber wait for 1st emit
+			distinctUntilChanged((curr, prev) => JSON.stringify(curr) === JSON.stringify(prev)),
+			map(source => ({
+				forum: {
+					[STORE.react]: source.filter(row => row[FIELD.store] === STORE.react) as IReact[],
+					[STORE.comment]: source.filter(row => row[FIELD.store] === STORE.comment) as IComment[],
+				}
+			})),
+		)
+	}
+
+	/**
 	 * Assemble an Object describing the Timetable for a specified date, as , where keys are:  
 	 * schedule   -> has an array of Schedule info for the weekday of the date  
 	 * calendar		-> has the Calendar event for that date (to override regular class)  
@@ -222,28 +263,26 @@ export class StateService {
 	 * location   -> has the Locations that are indicated on that schedule or calendar
 	 * instructor -> has the Instructors that are indicated on that schedule or calendar
 	 * span				-> has the Class Duration definitions ('full' or 'half')
-	 * feedback		-> has an array of Member Feedback about the Classes available on that schedule or event
+	 * forum			-> has an array of Member comments about the Classes available on that schedule or event
 	 * alert			-> has an array of Alert notes to display on the Attend component
 	 * bonus			-> has an array of active Bonus schemes
 	 * gift				-> has an array of active Gifts available to a Member on the date
+	 * forum			-> has an array of Forum data for the date
 	 */
-	getScheduleData(date?: TDate, elect?: BONUS) {
+	getScheduleData(date?: TDate, elect?: BONUS): Observable<ITimetableState> {
 		const now = getDate(date);
 		const isMine = addWhere(FIELD.uid, `{{auth.current.uid}}`);
-		const noToday = addWhere(`track.${FIELD.date}`, now.format(DATE_FMT.yearMonthDay), '!=');
+		const notToday = addWhere(`track.${FIELD.date}`, now.format(Instant.FORMAT.yearMonthDay), '!=');
 
 		const filterSchedule = addWhere('day', now.dow);
-		const filterCalendar = addWhere(FIELD.key, now.format(DATE_FMT.yearMonthDay));
+		const filterCalendar = addWhere(FIELD.key, now.format(Instant.FORMAT.yearMonthDay));
 		const filterEvent = addWhere(FIELD.key, `{{client.calendar.${FIELD.type}}}`);
 		const filterTypeClass = addWhere(FIELD.key, `{{client.schedule.${FIELD.key}}}`);
-		const filterTypeEvent = addWhere(FIELD.key, `{{client.event.classes}}`);
+		const filterTypeEvent = addWhere(FIELD.key, `{{client.event.agenda}}`);
 		const filterLocation = addWhere(FIELD.key, ['{{client.schedule.location}}', '{{client.calendar.location}}']);
 		const filterInstructor = addWhere(FIELD.key, ['{{client.schedule.instructor}}', '{{client.calendar.instructor}}']);
 		const filterSpan = addWhere(FIELD.key, [`{{client.class.${FIELD.type}}}`, `{{client.class.${FIELD.key}}}`]);
-		const filterFeedback = [
-			addWhere(FIELD.type, STORE.attend),
-			addWhere(FIELD.key, `{{client.schedule.${FIELD.key}}}`),
-		]
+		const filterIcon = addWhere(FIELD.type, [STORE.class, STORE.default]);
 		const filterAlert = [
 			addWhere(FIELD.type, STORE.schedule),
 			addWhere('location', ['{{client.schedule.location}}', '{{client.calendar.location}}']),
@@ -253,17 +292,20 @@ export class StateService {
 			addWhere(FIELD.expire, Number.MAX_SAFE_INTEGER, '<'),
 		]
 		const attendGift = [isMine, addWhere(`bonus.${FIELD.id}`, `{{member.gift[*].${FIELD.id}}}`)];
-		const attendWeek = [isMine, noToday, addWhere('track.week', now.format(DATE_FMT.yearWeek))];
-		const attendMonth = [isMine, noToday, addWhere('track.month', now.format(DATE_FMT.yearMonth))];
-		const attendToday = [isMine, addWhere(`track.${FIELD.date}`, now.format(DATE_FMT.yearMonthDay))];
+		const attendWeek = [isMine, notToday, addWhere('track.week', now.format(Instant.FORMAT.yearWeek))];
+		const attendMonth = [isMine, notToday, addWhere('track.month', now.format(Instant.FORMAT.yearMonth))];
+		const attendToday = [isMine, addWhere(`track.${FIELD.date}`, now.format(Instant.FORMAT.yearMonthDay))];
 
-		return this.getMemberData(date).pipe(
+		return combineLatest(this.getForumData(date), this.getMemberData(date)).pipe(
+			map(([forum, member]) => ({ ...forum, ...member })),
+			joinDoc(this.states, 'application', STORE.default, addWhere(FIELD.type, STORE.icon)),
 			joinDoc(this.states, 'client', STORE.schedule, filterSchedule, date),								// whats on this weekday
 			joinDoc(this.states, 'client', STORE.calendar, filterCalendar, date, calendarDay),	// get calendar for this date
 			joinDoc(this.states, 'client', STORE.calendar, filterRange, date),									// get any blocked calendar-range
 			joinDoc(this.states, 'client', STORE.event, filterEvent, date),											// get event for this calendar-date
 			joinDoc(this.states, 'client', STORE.class, filterTypeClass, date),									// get classes for this weekday
-			joinDoc(this.states, 'client', STORE.class, filterTypeEvent, date),									// get classes for this calendar-date
+			joinDoc(this.states, 'client', STORE.class, filterTypeEvent, date),									// get agenda for this calendar-date
+			joinDoc(this.states, 'client', STORE.icon, filterIcon, date),												// get the Class icons
 			joinDoc(this.states, 'client', STORE.location, filterLocation, date),								// get location for this timetable
 			joinDoc(this.states, 'client', STORE.instructor, filterInstructor, date),						// get instructor for this timetable
 			joinDoc(this.states, 'client', STORE.span, filterSpan, date),												// get class durations
@@ -271,9 +313,9 @@ export class StateService {
 			joinDoc(this.states, 'client', STORE.bonus, undefined, date),												// get any active Bonus
 			joinDoc(this.states, 'member', STORE.gift, isMine, date),														// get any active Gifts
 			joinDoc(this.states, 'attend.attendGift', STORE.attend, attendGift),								// get any Attends against active Gifts
-			joinDoc(this.states, 'attend.attendWeek', STORE.attend, attendWeek),								// get any Attends against this week
-			joinDoc(this.states, 'attend.attendMonth', STORE.attend, attendMonth),							// get any Attends against this month
-			joinDoc(this.states, 'attend.attendToday', STORE.attend, attendToday),							// get any Attends against this day
+			joinDoc(this.states, 'attend.attendWeek', STORE.attend, attendWeek),								// get any Attends for this week
+			joinDoc(this.states, 'attend.attendMonth', STORE.attend, attendMonth),							// get any Attends for this month
+			joinDoc(this.states, 'attend.attendToday', STORE.attend, attendToday),							// get any Attends for this day)
 			map(table => buildTimetable(table, date, elect)),																		// assemble the Timetable
 		) as Observable<ITimetableState>																											// declare Type (to override pipe()'s artificial limit of 'nine' declarations)
 	}
@@ -287,8 +329,8 @@ export class StateService {
 		const filterClass = addWhere(FIELD.key, `{{client.schedule.${FIELD.key}}}`);
 		const filterLocation = addWhere(FIELD.key, '{{client.schedule.location}}');
 		const filterCalendar = [
-			addWhere(FIELD.key, now.startOf('week').format(DATE_FMT.yearMonthDay), '>='),
-			addWhere(FIELD.key, now.endOf('week').format(DATE_FMT.yearMonthDay), '<='),
+			addWhere(FIELD.key, now.startOf('week').format(Instant.FORMAT.yearMonthDay), '>='),
+			addWhere(FIELD.key, now.endOf('week').format(Instant.FORMAT.yearMonthDay), '<='),
 		]
 		const filterEvent = addWhere(FIELD.key, `{{client.calendar.${FIELD.type}}}`);
 
