@@ -1,46 +1,44 @@
-import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Component, OnDestroy } from '@angular/core';
+import { Observable, Subscription, of } from 'rxjs';
+import { map, delay } from 'rxjs/operators';
 
-import { DialogService } from '@service/material/dialog.service';
+import { ForumService } from '@service/forum/forum.service';
 import { AttendService } from '@service/member/attend.service';
+import { DialogService } from '@service/material/dialog.service';
 
 import { ITimetableState } from '@dbase/state/state.define';
 import { StateService } from '@dbase/state/state.service';
-import { FIELD, COLLECTION } from '@dbase/data/data.define';
-import { IReact } from '@dbase/data/data.schema';
+import { FIELD, REACT } from '@dbase/data/data.define';
 import { DataService } from '@dbase/data/data.service';
 
-import { isUndefined } from '@lib/type.library';
-import { Instant } from '@lib/date.library';
+import { isUndefined, TString } from '@lib/type.library';
+import { Instant } from '@lib/instant.library';
 import { suffix } from '@lib/number.library';
 import { swipe } from '@lib/html.library';
 import { dbg } from '@lib/logger.library';
+import { ISchedule, IForumBase } from '@dbase/data/data.schema';
 
 @Component({
 	selector: 'wb-attend',
 	templateUrl: './attend.component.html',
+	styleUrls: ['./attend.component.scss'],
 })
-export class AttendComponent implements OnInit {
+export class AttendComponent implements OnDestroy {
 	private dbg = dbg(this);
-	private date!: number;
+	public date!: Instant;															// the date for the Schedule to display
+	public offset!: number;															// the number of days before today 
+	public firstPaint = true;                           // indicate first-paint
 
 	public selectedIndex: number = 0;                   // used by UI to swipe between <tabs>
 	public locations: number = 0;                       // used by UI to swipe between <tabs>
-	public timetable$!: Observable<ITimetableState>;
-	public firstPaint = true;                           // indicate first-paint
+	public timetable$!: Observable<ITimetableState>;		// the date's Schedule
+	private timerSubscription!: Subscription;						// watch for midnight, then reset this.date
 
 	constructor(private readonly attend: AttendService, public readonly state: StateService,
-		public readonly data: DataService, private dialog: DialogService) { }
+		public readonly data: DataService, private dialog: DialogService, private forum: ForumService) { this.setDate(0); }
 
-	ngOnInit() {                                        // wire-up the timetable Observable
-		this.timetable$ = this.state.getScheduleData(this.date).pipe(
-			map(data => {
-				this.locations = (data.client.location || []).length;
-				this.selectedIndex = 0;                       // start on the first-page
-				return data;
-			})
-		)
+	ngOnDestroy() {
+		this.timerSubscription && this.timerSubscription.unsubscribe();
 	}
 
 	// Build info to show in a Dialog
@@ -53,26 +51,26 @@ export class AttendComponent implements OnInit {
 		const dow = Instant.WEEKDAY[item.day];
 		// const bonus = client.bonus!.find(row => row[FIELD.key] === item.bonus);
 
-		const title = item.key;
+		const title = item[FIELD.key];
 		const subtitle = event && event.desc || '';
-		const icon = item.icon;
+		const image = item.image;
 		const actions = ['Close'];
 
 		let content: string[] = [];
 		if (!isUndefined(item.amount))
 			content.push(`Cost: $${item.amount.toFixed(2)}`);
 		if (item.start)
-			content.push(`Start: ${dow} @ ${item.start}`)
+			content.push(`Start: ${dow} @ ${new Instant(item.start).format(Instant.FORMAT.hhmi)}`)
 		if (span)
 			content.push(`Duration: ${span.duration} minutes`);
 		if (item.count)
-			content.push(`Attended: ${item.count} times today`);
+			content.push(`Attended: ${item.count} time${item.count > 1 ? 's' : ''}`);
 		if (locn)
 			content.push(`Location: ${locn.name}`);
 		if (instr)
 			content.push(`Instructor: ${instr.name}`);
 
-		this.dialog.open({ content, title, subtitle, icon, actions });
+		this.dialog.open({ content, title, subtitle, image, actions });
 	}
 
 	swipe(idx: number, event: any) {
@@ -82,5 +80,66 @@ export class AttendComponent implements OnInit {
 
 	suffix(idx: number) {
 		return suffix(idx);
+	}
+
+	/**
+	 * Call this onInit, whenever the Member changes the UI-date, and at midnight.   
+	 * It will allow a Member to check-in to a Class up-to 6 days in the past,  
+	 * useful if they forgot to scan at the time.  
+	 * dir:	if -1,	show previous day  
+	 * 			if 1,		show next day  
+	 * 			if 0,		show today
+	 */
+	setDate(dir: -1 | 0 | 1) {
+		const today = new Instant();
+		const offset = dir === 0
+			? today
+			: new Instant(this.date).add(dir, 'days')
+		this.offset = today.diff('days', offset);
+
+		this.date = this.offset > 6			// only allow up-to 6 days in the past
+			? today												// else reset to today
+			: offset
+
+		this.getSchedule();							// get day's Schedule
+
+		if (!this.timerSubscription)
+			this.setTimer();							// set a Schedule-view timeout
+	}
+
+	private getSchedule() {
+		this.timetable$ = this.state.getScheduleData(this.date).pipe(
+			map(data => {
+				this.selectedIndex = 0;                       // start on the first-page
+				return data;
+			})
+		)
+	}
+
+	/** If the Member is still sitting on this page at midnight, move this.date to next day */
+	private setTimer() {
+		const defer = new Instant().add(1, 'day').startOf('day');
+		this.dbg('timeOut: %s', defer.format('ddd, yyyy-mmm-dd HH:MI'));
+
+		this.timerSubscription && this.timerSubscription.unsubscribe();
+		this.timerSubscription = of(0)										// a single-emit Observable
+			.pipe(delay(defer.toDate()))
+			.subscribe(() => {
+				this.timerSubscription.unsubscribe();					// stop watching for midnight
+				this.setDate(0);															// onNext, show new day's timetable
+			})
+	}
+
+	public getForum() {
+		this.forum.getForum<IForumBase>()
+			.then(forum => this.dbg('forum: %j', forum))
+	}
+	public setReact(item: ISchedule, react: REACT) {
+		this.forum.setReact({ key: item[FIELD.id], track: { class: item[FIELD.key] }, react })
+			.then(_ => this.getForum())
+	}
+	public setComment(item: ISchedule, comment: TString) {
+		this.forum.setComment({ key: item[FIELD.id], track: { class: item[FIELD.key] }, comment })
+			.then(_ => this.getForum())
 	}
 }
