@@ -7,8 +7,11 @@ export function Sealed(constructor: Function) {
 	Object.seal(constructor.prototype);
 }
 
-export function Loggable<T extends { new(...args: any[]): {} }>(constructor: T) {
+export function Debug<T extends { new(...args: any[]): {} }>(constructor: T) {
 	return class extends constructor {
 		private dbg = dbg(this);
+		constructor(...args: any[]) {
+			super(...args);
+		}
 	}
 }
