@@ -1,6 +1,6 @@
 import { fix } from '@lib/number.library';
 import { asString, asNumber } from '@lib/string.library';
-import { isString, isNumber, getType, isUndefined } from '@lib/type.library';
+import { getType, isString, isNumber, isUndefined } from '@lib/type.library';
 
 interface IInstant {											// Date components
 	readonly yy: number;										// year[4]
@@ -15,39 +15,30 @@ interface IInstant {											// Date components
 	readonly mmm: string;										// short month-name
 	readonly ddd: string;										// short day-name
 	readonly dow: number;										// weekday; Mon=1, Sun=7
+	readonly tz: number;										// timezone offset in hours
 	readonly value?: TDate;									// original value passed to constructor
 }
 
 interface ITimestamp {
-	seconds: number;
-	nanoseconds: number;
-};
+	readonly seconds: number;
+	readonly nanoseconds: number;
+}
 
 interface IDateFmt {											// pre-configured format strings
 	[str: string]: string | number;					// allow for dynamic format-codes
-	"ddd, dd mmm yyyy": string;
-	"yyyy-mm-dd HH:MI": string;
-	"dd-mmm": string;
-	"HH:MI": string;
-	"yyyyww": number;
-	"yyyymm": number;
-	"yyyymmdd": number;
-}
-
-export enum DATE_FMT {										// pre-configured format names
-	display = 'ddd, dd mmm yyyy',
-	dateTime = 'yyyy-mm-dd HH:MI',
-	dayMonth = 'dd-mmm',
-	HHMI = 'HH:MI',
-	yearWeek = 'yyyyww',
-	yearMonth = 'yyyymm',
-	yearMonthDay = 'yyyymmdd',
+	readonly "ddd, dd mmm yyyy": string;
+	readonly "yyyy-mm-dd HH:MI": string;
+	readonly "dd-mmm": string;
+	readonly "HH:MI": string;
+	readonly "hh:mi": string;
+	readonly "yyyyww": number;
+	readonly "yyyymm": number;
+	readonly "yyyymmdd": number;
 }
 
 type TArgs = string[] | number[];
 type TMutate = 'add' | 'start' | 'mid' | 'end';
-type TUnitTime = 'year' | 'years' | 'month' | 'months' | 'day' | 'days' | 'hour' | 'hours' | 'minute' | 'minutes' | 'week' | 'weeks';
-type TUnitOffset = 'month' | 'week' | 'day';
+type TUnitTime = 'year' | 'month' | 'week' | 'day' | 'hour' | 'minute' | 'second';
 type TUnitDiff = 'years' | 'months' | 'weeks' | 'days' | 'hours' | 'minutes' | 'seconds';
 
 export type TDate = string | number | Date | Instant | IInstant | Timestamp | ITimestamp;
@@ -56,6 +47,7 @@ export type TDate = string | number | Date | Instant | IInstant | Timestamp | IT
 class Timestamp {
 	readonly seconds: number;
 	readonly nanoseconds: number;
+	private readonly now: number;
 
 	constructor(seconds?: number, nanoseconds?: number) {
 		const now = new Date();
@@ -65,9 +57,10 @@ class Timestamp {
 		this.nanoseconds = isUndefined(nanoseconds)
 			? now.getMilliseconds() * 1000000
 			: nanoseconds
+		this.now = now.setHours(0, 0, this.seconds, this.nanoseconds / 1000);
 	}
 
-	public toDate = () => new Date((this.seconds * 1000) + Math.floor(this.nanoseconds / 1000000));
+	public toDate = () => new Date(this.now);
 	static fromDate = (dt: Date) => new Timestamp(dt.getSeconds(), dt.getMilliseconds() * 1000000);
 	static fromStamp = (stamp: number) => Timestamp.fromDate(new Date(stamp * 1000));
 	static now = new Timestamp();
@@ -77,8 +70,10 @@ class Timestamp {
 
 // shortcut functions to common Instant class properties / methods.
 /** get new Instant */export const getDate = (dt?: TDate, ...args: TArgs) => new Instant(dt, ...args);
-/** get Timestamp */	export const getStamp = (dt?: TDate, ...args: TArgs) => getDate(dt, ...args).ts;
+/** get timestamp */	export const getStamp = (dt?: TDate, ...args: TArgs) => getDate(dt, ...args).ts;
 /** format Instant */	export const fmtDate = <K extends keyof IDateFmt>(fmt: K, dt?: TDate, ...args: TArgs) => getDate(dt, ...args).format(fmt);
+
+// NOTE: Instant does not currently handle leap-seconds
 
 /**
  * An Instant is a object that is used to manage Javascript Dates.  
@@ -88,6 +83,23 @@ class Timestamp {
  */
 export class Instant {
 	private date: IInstant;																				// Date parsed into components
+
+	private static PATTERN = {
+		hhmi: /^\d\d:\d\d$/,																				// regex to match HH:MI
+		yyyymmdd: /^(19\d{2}|20\d{2})(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$/,
+		ddmmyyyy: /^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[012])\/(19\d{2}|20\d{2})$/,
+	}
+	private static maxStamp = new Date('9999-12-31').valueOf() / 1000;
+	private static minStamp = new Date('1000-01-01').valueOf() / 1000;
+	private static divideBy = {																		// approx date-offset divisors (unix-timestamp precision)
+		years: 31536000000,
+		months: 2628000000,
+		weeks: 604800000,
+		days: 86400000,
+		hours: 3600000,
+		minutes: 60000,
+		seconds: 1000,
+	}
 
 	constructor(dt?: TDate, ...args: TArgs) { this.date = this.parseDate(dt, args); }
 
@@ -100,6 +112,7 @@ export class Instant {
 	/** seconds */				get SS() { return this.date.SS }
 	/** timestamp */			get ts() { return this.date.ts }
 	/** milliseconds */		get ms() { return this.date.ms }
+	/** timezone offset */get tz() { return this.date.tz }
 	/** number of weeks*/	get ww() { return this.date.ww }
 	/** weekday number */	get dow() { return this.date.dow }
 	/** short day name */	get ddd() { return this.date.ddd }
@@ -111,28 +124,28 @@ export class Instant {
 	/** calc diff Dates, default as <years> */
 	diff = (unit: TUnitDiff = 'years', dt2?: TDate, ...args: TArgs) => this.diffDate(unit, dt2, ...args);
 	/** add date offset, default as <minutes> */
-	add = (offset: number, unit: TUnitTime = 'minutes') => this.setDate('add', unit, offset);
+	add = (offset: number, unit: TUnitTime | TUnitDiff = 'minutes') => this.setDate('add', unit, offset);
 
 	/** start offset, default as <week> */
-	startOf = (unit: TUnitOffset = 'week') => this.setDate('start', unit);
+	startOf = (unit: TUnitTime = 'week') => this.setDate('start', unit);
 	/** middle offset, default as <week> */
-	midOf = (unit: TUnitOffset = 'week') => this.setDate('mid', unit);
+	midOf = (unit: TUnitTime = 'week') => this.setDate('mid', unit);
 	/** ending offset, default as <week> */
-	endOf = (unit: TUnitOffset = 'week') => this.setDate('end', unit);
+	endOf = (unit: TUnitTime = 'week') => this.setDate('end', unit);
 
 	/** valid Instant */	isValid = () => !isNaN(this.date.ts);
-	/** get raw object */	asObject = () => ({ ...this.date });
-	/** as Date object */	asDate = () => new Date(this.date.ts * 1000 + this.date.ms);
+	/** get raw object */	toJSON = () => ({ ...this.date });
+	/** as Date object */	toDate = () => this.composeDate(this.date);
 
 	// Private methods	~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 	/** parse a Date, return components */
 	private parseDate = (dt?: TDate, args: TArgs = []) => {
 		let date: Date;
 
-		if (isString(dt) && Instant.hhmi.test(dt))								// if only HH:MI supplied...
+		if (isString(dt) && Instant.PATTERN.hhmi.test(dt))								// if only HH:MI supplied...
 			dt = `${new Date().toDateString()} ${dt}`;							// 	prepend current date
-		if ((isString(dt) || isNumber(dt)) && Instant.yyyymmdd.test(asString(dt)))// if yyyymmdd supplied as string...
-			dt = asString(dt).replace(Instant.yyyymmdd, '$1-$2-$3 00:00:00');				// format to look like a date-string
+		if ((isString(dt) || isNumber(dt)) && Instant.PATTERN.yyyymmdd.test(asString(dt)))// if yyyymmdd supplied as string...
+			dt = asString(dt).replace(Instant.PATTERN.yyyymmdd, '$1-$2-$3 00:00:00');				// format to look like a date-string
 
 		switch (getType(dt)) {																		// convert 'dt' argument into a Date
 			case 'Undefined':																				// default to 'now'
@@ -140,14 +153,24 @@ export class Instant {
 				break;
 
 			case 'Date':																						// already is a valid Date	
-				date = dt as Date;
+				date = new Date(dt as Date);													// clone a new Date
 				break;
 
 			case 'String':
+				if (Instant.WEEKDAY[dt as keyof typeof Instant.WEEKDAY]) {
+					const now = new Date();															// 3-character Weekday
+					now.setDate(now.getDate() - 7 - now.getDay() + Instant.WEEKDAY[dt as keyof typeof Instant.WEEKDAY]);
+					dt = `${now.getFullYear()}-${now.getMonth() + 1}-${now.getDate()}`
+				}
+				if (Instant.WEEKDAYS[dt as keyof typeof Instant.WEEKDAYS]) {
+					const now = new Date();															// full character Weekday
+					now.setDate(now.getDate() - 7 - now.getDay() + Instant.WEEKDAYS[dt as keyof typeof Instant.WEEKDAYS]);
+					dt = `${now.getFullYear()}-${now.getMonth() + 1}-${now.getDate()}`
+				}
 				const fmt = dt as string;
-				date = new Date(args.length
+				date = args.length
 					? this.formatString(fmt, ...args)										// free-format string
-					: fmt)																							// let Javascript interpret string
+					: new Date(fmt)																			// let browser interpret string
 
 				if (date.getFullYear() === 2001) {										// Javascript default year, if none in parse-string
 					if (!fmt.includes('2001'))													// did not specifically request default year
@@ -164,7 +187,7 @@ export class Instant {
 				break;
 
 			case 'Instant':																					// already have a valid Instant
-				date = (dt as Instant).asDate();
+				date = (dt as Instant).toDate();
 				break;
 
 			case 'Timestamp':																				// instance of Timestamp
@@ -190,24 +213,28 @@ export class Instant {
 		if (isNaN(date.getTime()))																// Date not parse-able,
 			console.error('Invalid Date: ', dt, date);							// TODO: log the Invalid Date
 
-		let [yy, mm, dd, HH, MI, SS, ts, ms, dow] = [
+		let [yy, mm, dd, HH, MI, SS, ts, ms, tz, dow] = [
 			date.getFullYear(), date.getMonth(), date.getDate(),
 			date.getHours(), date.getMinutes(), date.getSeconds(),
-			Math.floor(date.getTime() / 1000), date.getMilliseconds(), date.getDay(),
+			Math.floor(date.getTime() / 1000), date.getMilliseconds(),
+			date.getTimezoneOffset(), date.getDay(),
 		];
 
 		mm += 1;																									// ISO month
-		if (!dow && !isNaN(dow)) dow = Instant.DAY.Sun;						// ISO weekday
+		if (!dow && !isNaN(dow)) dow = Instant.WEEKDAY.Sun;				// ISO weekday
 
-		const thu = date.setDate(dd - dow + Instant.DAY.Thu);			// set to nearest Thursday
+		const thu = date.setDate(dd - dow + Instant.WEEKDAY.Thu);	// set to nearest Thursday
 		const ny = new Date(date.getFullYear(), 0, 1).valueOf();	// NewYears Day
 		const ww = Math.floor((thu - ny) / Instant.divideBy.weeks + 1);	// ISO Week Number
 
-		return { yy, mm, dd, HH, MI, SS, ts, ms, ww, dow, ddd: Instant.DAY[dow], mmm: Instant.MONTH[mm], value: dt } as IInstant;
+		return {
+			yy, mm, dd, HH, MI, SS, ts, ms, tz, ww, dow,
+			ddd: Instant.WEEKDAY[dow], mmm: Instant.MONTH[mm], value: dt
+		} as IInstant;
 	}
 
-	/** mutate an Instant */
-	private setDate = (mutate: TMutate, unit: TUnitTime, offset: number = 1) => {
+	/** create a new offset Instant */
+	private setDate = (mutate: TMutate, unit: TUnitTime | TUnitDiff, offset: number = 1) => {
 		const date = { ...this.date };														// clone the current Instant
 		const single = unit.endsWith('s')
 			? unit.substring(0, unit.length - 1)										// remove plural units
@@ -216,57 +243,80 @@ export class Instant {
 			[date.HH, date.MI, date.SS, date.ms] = [0, 0, 0, 0];		// discard the time-portion of the date
 
 		switch (`${mutate}.${single}`) {
-			case 'start.minute':
-				[date.SS, date.ms] = [0, 0];
+			case 'start.year':
+				[date.mm, date.dd] = [1, 1];
+			case 'start.month':
+				date.dd = 1;
 				break;
-			case 'start.hour':
-				[date.MI, date.SS, date.ms] = [0, 0, 0];
+			case 'start.week':
+				date.dd -= date.dow + Instant.WEEKDAY.Mon;
 				break;
 			case 'start.day':
 				[date.HH, date.MI, date.SS, date.ms] = [0, 0, 0, 0];
 				break;
-			case 'start.week':
-				date.dd -= date.dow + Instant.DAY.Mon;
+			case 'start.hour':
+				[date.MI, date.SS, date.ms] = [0, 0, 0];
 				break;
-			case 'start.month':
-				date.dd = 1;
+			case 'start.minute':
+				[date.SS, date.ms] = [0, 0];
 				break;
-			case 'mid.second':
-				[date.SS, date.ms] = [30, 0];
+			case 'start.second':
+				date.ms = 0;
 				break;
-			case 'mid.minute':
-				[date.MI, date.SS, date.ms] = [30, 0, 0];
-				break;
-			case 'mid.day':
-				[date.HH, date.MI, date.SS, date.ms] = [12, 0, 0, 0];
-				break;
-			case 'mid.week':
-				date.dd -= date.dow + Instant.DAY.Thu;
+
+			case 'mid.year':
+				[date.mm, date.dd] = [7, 1];
 				break;
 			case 'mid.month':
 				date.dd = 15;
 				break;
-			case 'end.second':
-				[date.SS, date.ms] = [59, 999];
+			case 'mid.week':
+				date.dd -= date.dow + Instant.WEEKDAY.Thu;
 				break;
-			case 'end.minute':
-				[date.MI, date.SS, date.ms] = [59, 59, 999];
+			case 'mid.day':
+				[date.HH, date.MI, date.SS, date.ms] = [12, 0, 0, 0];
 				break;
-			case 'end.day':
-				[date.HH, date.MI, date.SS, date.ms] = [23, 59, 59, 999];
+			case 'mid.hour':
+				[date.MI, date.SS, date.ms] = [30, 0, 0];
 				break;
-			case 'end.week':
-				date.dd -= date.dow + Instant.DAY.Sun;
+			case 'mid.minute':
+				[date.SS, date.ms] = [30, 0];
+				break;
+			case 'mid.second':
+				date.ms = 500;
+				break;
+
+			case 'end.year':
+				[date.mm, date.dd] = [12, 31];
 				break;
 			case 'end.month':
 				date.mm += 1;
 				date.dd = 0;
 				break;
+			case 'end.week':
+				date.dd -= date.dow + Instant.WEEKDAY.Sun;
+				break;
+			case 'end.day':
+				[date.HH, date.MI, date.SS, date.ms] = [23, 59, 59, 999];
+				break;
+			case 'end.hour':
+				[date.MI, date.SS, date.ms] = [59, 59, 999];
+				break;
+			case 'end.minute':
+				[date.MI, date.SS, date.ms] = [59, 59, 999];
+				break;
+			case 'end.second':
+				[date.SS, date.ms] = [59, 999];
+				break;
+
 			case 'add.year':
 				date.yy += offset;
 				break;
 			case 'add.month':
 				date.mm += offset;
+				break;
+			case 'add.week':
+				date.dd += offset * 7;
 				break;
 			case 'add.day':
 				date.dd += offset;
@@ -277,8 +327,8 @@ export class Instant {
 			case 'add.minute':
 				date.MI += offset;
 				break;
-			case 'add.week':
-				date.dd += offset * 7;
+			case 'add.second':
+				date.SS += offset;
 				break;
 		}
 
@@ -293,17 +343,19 @@ export class Instant {
 		const date = { ...this.date };													// clone current Instant
 
 		switch (fmt) {
-			case DATE_FMT.yearWeek:
+			case Instant.FORMAT.yearWeek:
 				const offset = date.ww === 1 && date.mm === 12;			// if late-Dec, add 1 to yy
 				return parseInt(`${date.yy + Number(offset)}${fix(date.ww)}`);
 
-			case DATE_FMT.yearMonth:
+			case Instant.FORMAT.yearMonth:
 				return parseInt(`${fix(date.yy)}${fix(date.mm)}`);
 
-			case DATE_FMT.yearMonthDay:
+			case Instant.FORMAT.yearMonthDay:
 				return parseInt(`${date.yy}${fix(date.mm)}${fix(date.dd)}`, 10);
 
 			default:
+				const am = date.HH >= 12 ? 'pm' : 'am';							// noon is considered 'pm'
+
 				return asString(fmt)
 					.replace(/y{4}/g, fix(date.yy))
 					.replace(/y{2}/g, fix(date.yy).substring(2, 4))
@@ -314,6 +366,9 @@ export class Instant {
 					.replace(/H{2}/g, fix(date.HH))
 					.replace(/M{2}/g, fix(date.MI))
 					.replace(/MI/g, fix(date.MI))
+					.replace(/h{2}/g, fix(date.HH >= 13 ? date.HH % 12 : date.HH))
+					.replace(/m{2}/g, fix(date.MI) + am)
+					.replace(/mi/g, fix(date.MI) + am)
 					.replace(/S{2}/g, fix(date.SS))
 					.replace(/ts/g, asString(date.ts))
 					.replace(/ms/g, asString(date.ms))
@@ -329,15 +384,15 @@ export class Instant {
 	 * e.g. ('%dd-%mmm', 20, 'May') will return new Date() for 20-May in current year.
 	 */
 	private formatString = (fmt: string, ...args: TArgs) => {
-		const date = new Instant().startOf('day').asObject()					// date components
-		let cnt = 0;
+		const date = new Instant().startOf('day').toJSON()					// date components
+		let argCnt = 0;
 
 		fmt.split('%')
 			.forEach(word => {
 				const word4 = word.substring(0, 4),
 					word3 = word.substring(0, 3),
 					word2 = word.substring(0, 2);
-				let param = args[cnt];
+				let param = args[argCnt];
 
 				switch (true) {
 					case word2 === 'yy' && word4 !== 'yyyy':
@@ -349,13 +404,13 @@ export class Instant {
 						break;
 
 					case word3 === 'mmm':						// change the month-name into a month-number
-						param = Instant.MONTH[param as keyof typeof Instant.MONTH];
+						param = Instant.MONTH[param as Instant.MONTH];
 					case word2 === 'mm':
 						date.mm = asNumber(param);
 						break;
 
 					case word3 === 'ddd':						// set the day to the last occurrence of %ddd
-						let weekDay = Instant.DAY[param as keyof typeof Instant.DAY];
+						let weekDay = Instant.WEEKDAY[param as Instant.WEEKDAY];
 						param = (date.dd - date.dow + weekDay).toString();
 					case word2 === 'dd':
 						date.dd = asNumber(param);
@@ -371,40 +426,40 @@ export class Instant {
 					case word2 === 'SS':
 						date.SS = asNumber(param)
 						break;
+
 					default:
-						cnt -= 1;
+						argCnt -= 1;
 				}
-				cnt += 1;
+				argCnt += 1;
 			})
 
 		return this.composeDate(date);
 	}
 
-	/** calculate the difference between dates */
+	/** calculate the difference between dates (past is positive, future is negative) */
 	private diffDate = (unit: TUnitDiff = 'years', dt2?: TDate, ...args: TArgs) => {
 		const offset = this.parseDate(dt2, args);
-		return Math.floor(((offset.ts * 1000 + offset.ms) - (this.date.ts * 1000 + this.date.ms)) / Instant.divideBy[unit]);
+		const diff = (this.date.ts * 1000 + this.date.ms) - (offset.ts * 1000 + offset.ms);
+		return diff < 0
+			? Math.ceil(diff / Instant.divideBy[unit])
+			: Math.floor(diff / Instant.divideBy[unit])
 	}
 }
 
 export namespace Instant {
-	export enum DAY { Mon = 1, Tue, Wed, Thu, Fri, Sat, Sun }
-	export enum WEEKDAY { Monday = 1, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday }
+	export enum WEEKDAY { Mon = 1, Tue, Wed, Thu, Fri, Sat, Sun }
+	export enum WEEKDAYS { Monday = 1, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday }
 	export enum MONTH { Jan = 1, Feb, Mar, Apr, May, Jun, Jul, Aug, Sep, Oct, Nov, Dec }
 	export enum MONTHS { January = 1, February, March, April, May, June, July, August, September, October, November, December }
 
-	export const hhmi = /^\d\d:\d\d$/;								// regex to match HH:MI
-	export const yyyymmdd = /^(19\d{2}|20\d{2})(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$/;	// regex to match YYYYMMDD
-	export const ddmmyyyy = /^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[012])\/(19\d{2}|20\d{2})$/;
-	export const maxStamp = new Date('9999-12-31').valueOf() / 1000;
-	export const minStamp = new Date('1000-01-01').valueOf() / 1000;
-	export const divideBy = {													// approx date-offset divisors (unix-timestamp precision)
-		years: 31536000000,
-		months: 2628000000,
-		weeks: 604800000,
-		days: 86400000,
-		hours: 3600000,
-		minutes: 60000,
-		seconds: 1000,
+	export enum FORMAT {											// pre-configured format names
+		display = 'ddd, dd mmm yyyy',
+		dateTime = 'yyyy-mm-dd HH:MI',
+		dayMonth = 'dd-mmm',
+		HHMI = 'HH:MI',													// 24-hour format
+		hhmi = 'hh:mi',													// 12-hour format
+		yearWeek = 'yyyyww',
+		yearMonth = 'yyyymm',
+		yearMonthDay = 'yyyymmdd',
 	}
 }
