@@ -17,24 +17,25 @@ export const getType = (obj?: any): string => {
 }
 
 /** Type-Guards: return a boolean to test <obj> is of <type> */
-export const isType = (obj: any, type: string = 'Object'): boolean => getType(obj).toLowerCase() === type.toLowerCase();
+export const isType = (obj: unknown, type: string = 'Object'): boolean => getType(obj).toLowerCase() === type.toLowerCase();
 
-export const isString = (obj?: any): obj is string => isType(obj, 'String');
-export const isNumber = (obj?: any): obj is number => isType(obj, 'Number');
-export const isInteger = (obj?: any): obj is bigint => isType(obj, 'BigInt');
+export const isString = (obj?: unknown): obj is string => isType(obj, 'String');
+export const isNumber = (obj?: unknown): obj is number => isType(obj, 'Number');
+export const isInteger = (obj?: unknown): obj is bigint => isType(obj, 'BigInt');
+export const isBoolean = (obj?: unknown): obj is boolean => isType(obj, 'Boolean');
 export const isObject = <T>(obj?: T): obj is T => isType(obj, 'Object');
-export const isArray = (obj?: any): obj is any[] => isType(obj, 'Array');
-export const isNull = (obj?: any): obj is null => isType(obj, 'Null');
-export const isDate = (obj?: any): obj is Date => isType(obj, 'Date');
-export const isFunction = (obj?: any): obj is Function => isType(obj, 'Function') || isType(obj, 'AsyncFunction');
-export const isClass = (obj?: any): obj is Function => isType(obj, 'Class');
-export const isPromise = <T>(obj?: any): obj is Promise<T> => isType(obj, 'Promise');
-export const isUndefined = (obj?: any): obj is undefined => isType(obj, 'Undefined');
-export const isBlob = (obj?: any): obj is Blob => isType(obj, 'Blob');
+export const isArray = (obj?: unknown): obj is any[] => isType(obj, 'Array');
+export const isNull = (obj?: unknown): obj is null => isType(obj, 'Null');
+export const isDate = (obj?: unknown): obj is Date => isType(obj, 'Date');
+export const isFunction = (obj?: unknown): obj is Function => isType(obj, 'Function') || isType(obj, 'AsyncFunction');
+export const isClass = (obj?: unknown): obj is Function => isType(obj, 'Class');
+export const isPromise = <T>(obj?: unknown): obj is Promise<T> => isType(obj, 'Promise');
+export const isUndefined = (obj?: unknown): obj is undefined => isType(obj, 'Undefined');
+export const isBlob = (obj?: unknown): obj is Blob => isType(obj, 'Blob');
 
-export const nullToZero = (obj?: any) => isUndefined(obj) || isNull(obj) ? 0 : obj;
-export const nullToEmpty = (obj?: any) => isUndefined(obj) || isNull(obj) ? '' : obj;
-export const nullToValue = (obj?: any, value?: any) => isUndefined(obj) || isNull(obj) ? value : obj;
+export const nullToZero = (obj?: number) => isUndefined(obj) || isNull(obj) ? 0 : obj;
+export const nullToEmpty = (obj?: string) => isUndefined(obj) || isNull(obj) ? '' : obj;
+export const nullToValue = (obj?: unknown, value?: any) => isUndefined(obj) || isNull(obj) ? value : obj;
 
 export type TString = string | string[];
 export type TNumber = number | number[];
