@@ -1,27 +1,15 @@
 import { UserInfo, firestore } from 'firebase/app';
 
-import { FIELD, STORE, TYPE, CONNECT, BONUS, REACT, COLLECTION } from '@dbase/data/data.define';
+import { COLLECTION, STORE, FIELD, CLASS, EVENT, CONNECT, BONUS, REACT, Auth, PLAN, PRICE, PAYMENT, PROFILE, STATUS, SCHEDULE, MESSAGE, SPAN } from '@dbase/data/data.define';
 import { ISummary } from '@dbase/state/state.define';
 import { getSlice } from '@dbase/state/state.library';
 
 import { TString } from '@lib/type.library';
 
 type TStoreConfig = STORE.schema | STORE.config | STORE.default;
-type TStoreClient = STORE.class | STORE.event | STORE.price | STORE.plan | STORE.provider | STORE.schedule | STORE.calendar | STORE.location | STORE.instructor | STORE.bonus | STORE.span | STORE.alert | STORE.react;
-type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status | STORE.feedback;
-type TTypeDefault = TStoreClient | 'icon';
-
-type TSpan = 'full' | 'half';
-type TPayment = TYPE.topUp | 'hold' | 'credit' | 'debit';
-type TProfile = TYPE.plan | TYPE.claim | TYPE.info | TYPE.pref;
-type TStatus = 'account' | 'connect';
-type TSchedule = 'event' | 'class' | 'special';
-type TCalendar = 'event' | 'special';
-
-export type TPrice = TSpan | TYPE.topUp | 'hold' | 'expiry';
-export type TClass = 'AeroStep' | 'HiLo' | 'MultiStep' | 'SingleStep' | 'SmartStep' | 'StepBasic' | 'StepDown' | 'StepIn' | 'Zumba' | 'ZumbaStep';
-export type TPlan = 'member' | 'casual' | 'gratis' | 'student' | 'core' | 'intro';
-export type TProvider = 'identity' | 'oauth' | 'oidc' | 'email' | 'play' | 'phone' | 'anonymous';
+type TStoreClient = STORE.class | STORE.event | STORE.price | STORE.plan | STORE.provider | STORE.schedule | STORE.calendar | STORE.location | STORE.instructor | STORE.bonus | STORE.span | STORE.alert | STORE.icon;
+type TStoreUser = STORE.profile | STORE.payment | STORE.gift | STORE.message | STORE.migrate | STORE.attend | STORE.register | STORE.status;
+type TStoreForum = STORE.comment | STORE.react;
 
 export type FBoolean = boolean | undefined | firestore.FieldValue;
 export type FNumber = number | undefined | firestore.FieldValue;
@@ -30,7 +18,6 @@ export type FType<T> = Exclude<T, firestore.FieldValue>;
 
 // These are the meta- and common-fields for a standard document
 export interface IMeta {
-	[FIELD.id]: string;										// the id field on the remote database, generated by FireBase
 	[FIELD.create]?: number;							// the time when originally created
 	[FIELD.update]?: number;							// the time when last updated
 	[FIELD.access]?: number;							// the time when last accessed
@@ -39,49 +26,61 @@ export interface IMeta {
 	[FIELD.hidden]?: FBoolean;						// valid value, but not to be displayed to the client
 	[FIELD.disable]?: FBoolean;						// valid value, greyed-out to the client
 
-	[FIELD.store]: STORE;									// every document must declare the 'store' field
+	[FIELD.id]: string;										// the id field on the remote database, generated by FireBase
+	[FIELD.store]: STORE;									// each document must declare a 'store' field
 	[FIELD.type]?: string | number;				// an optional 'type' code to qualify the 'store'
-	[FIELD.note]?: FString;								// an optional 'note' on all documents
+	[FIELD.note]?: TString;								// an optional 'note' on all documents
 }
 
 /**
  * We have two main types of 'store' documents.  
  * One for 'client' & 'local', keyed by 'store/type?/key',  
- * the other for 'admin', 'member' & 'attend', keyed by 'store/type?/uid'
+ * the other for 'admin', 'member', 'forum' & 'attend', keyed by 'store/type?/uid'
  */
-export type TStoreBase = IClientBase | IUserBase | IMigrateBase;
+export type TStoreBase = IClientBase | IUserBase | IForumBase | IMigrate;
 interface IClientBase extends IMeta {
 	[FIELD.store]: TStoreClient | TStoreConfig;
 	[FIELD.key]: string | number;
-	[FIELD.icon]?: string;									// an optional icon for the UI
+	[FIELD.image]?: string;									// an optional icon for the UI
 }
 interface IUserBase extends IMeta {				// this is the base for Member-related documents
 	[FIELD.store]: TStoreUser;
 	[FIELD.uid]: string;
 	[FIELD.stamp]: number;
 }
-export interface IMigrateBase extends IMeta {
+export interface IForumBase extends IMeta {
+	[FIELD.store]: TStoreForum;
+	[FIELD.type]: STORE;										// allow for Forum on any Store type
+	[FIELD.uid]: string;										// the Member making the feedback
+	[FIELD.key]: string;										// key to the Store the Member is referencing
+	[FIELD.stamp]: number;
+	track: {																// to use in feedback-analysis
+		date: number;													// yearMonthDay
+		[key: string]: string | number;				// additional info to assist tracking
+	},
+}
+export interface IMigrate extends IMeta {	// these allow for reconciliation of migrated event
 	[FIELD.store]: STORE.migrate;
 	[FIELD.type]: STORE.event | STORE.class;
 	[FIELD.uid]: string;
 	[FIELD.key]: string;
 	attend: {
-		[order: string]: string;
+		[order: string]: CLASS;
 	}
 }
 export interface IStoreMeta extends IMeta {
-	store: STORE;
+	[FIELD.store]: STORE;
 	[key: string]: any;											// add in index-signature
 }
 export type TStoreMeta = IStoreMeta | IStoreMeta[];
 // client documents have a '<key>' field, user documents have a '<uid>' field
 export const isClientDocument = (document: TStoreBase): document is IClientBase =>
-	getSlice(document[FIELD.store]).toString() === 'client' || getSlice(document[FIELD.store]).toString() === 'local';
+	getSlice(document[FIELD.store]).toString() === COLLECTION.client || getSlice(document[FIELD.store]).toString() === STORE.local;
 
 //	/client/_default_
 export interface IDefault extends IClientBase {
 	[FIELD.store]: STORE.default;
-	[FIELD.type]: TTypeDefault;
+	[FIELD.type]: TStoreClient;						// Client store to default
 	[FIELD.key]: string;
 }
 //	/client/_config_
@@ -103,8 +102,8 @@ export interface ISchema extends IClientBase {
 //	/client/price
 export interface IPrice extends IClientBase {
 	[FIELD.store]: STORE.price;
-	[FIELD.type]: TPrice;
-	[FIELD.key]: TPlan;
+	[FIELD.type]: PRICE;
+	[FIELD.key]: PLAN;
 	amount: number;
 }
 
@@ -124,18 +123,19 @@ export interface IBonus extends IClientBase {
 //	/client/plan
 export interface IPlan extends IClientBase {
 	[FIELD.store]: STORE.price;
-	[FIELD.key]: string;
+	[FIELD.key]: PLAN;
 	bonus?: boolean;										// allowed to apply Bonus
 	desc: string;
 	sort: number;
 	rule?: TString;
 	expiry?: number;										// number of months a Payment is active (before it is auto-expired)
 }
+
 //	/client/class
 export interface IClass extends IClientBase {
 	[FIELD.store]: STORE.class;
-	[FIELD.type]: TSpan;
-	[FIELD.key]: TClass;
+	[FIELD.type]: SPAN;
+	[FIELD.key]: CLASS;
 	color: string;
 	desc?: string;
 }
@@ -144,15 +144,15 @@ export interface IClass extends IClientBase {
 export interface IEvent extends IClientBase {
 	[FIELD.store]: STORE.event;
 	[FIELD.key]: string;
+	agenda: CLASS[];									// the Classes on offer for this Event
 	name: string;
 	desc?: string;
-	classes: TClass[];
 }
 
 //	/client/calendar
 export interface ICalendar extends IClientBase {
 	[FIELD.store]: STORE.calendar;
-	[FIELD.type]: TCalendar;
+	[FIELD.type]: EVENT;
 	[FIELD.key]: number;
 	name: string;
 	desc?: string;
@@ -171,10 +171,14 @@ export interface TBonus {							// a sub-type of IBonus
 	gift?: IGift[];											// the /member/gift updates for this payment
 	amount?: number;										// an optional amount (if other than $0)
 }
+export type TForum = {
+	comment?: TString;
+	react?: REACT;
+}
 export interface ISchedule extends IClientBase {
 	[FIELD.store]: STORE.schedule | STORE.calendar;
-	[FIELD.type]: TSchedule;
-	[FIELD.key]: TClass | string;
+	[FIELD.type]: SCHEDULE;
+	[FIELD.key]: CLASS;
 	day: number;
 	location?: string;
 	instructor?: string;
@@ -185,7 +189,7 @@ export interface ISchedule extends IClientBase {
 	bonus?: TBonus;											// the Bonus tracking which can be applied to this Schedule
 	elect?: BONUS;											// name the Bonus the Member chooses (override calc)
 	count?: number;											// number of Attends
-	react?: REACT;											// Member reaction
+	forum?: TForum;
 }
 
 //	/client/location
@@ -226,7 +230,8 @@ export interface IInstructor extends IClientBase {
 //	/client/span
 export interface ISpan extends IClientBase {
 	[FIELD.store]: STORE.span;
-	[FIELD.key]: TSpan;
+	[FIELD.key]: SPAN | CLASS;
+	[FIELD.type]: SCHEDULE;
 	duration: number;
 }
 
@@ -234,25 +239,27 @@ export interface ISpan extends IClientBase {
 export interface IAlert extends IClientBase {
 	[FIELD.store]: STORE.alert;
 	[FIELD.type]: STORE.schedule | STORE.event;
-	[FIELD.note]: TString;
 	location?: string;
 }
 
 //	/client/react									// Member reaction icons
-export interface IReact extends IClientBase {
-	[FIELD.store]: STORE.react;
-	[FIELD.key]: REACT;
+export interface IIcon extends IClientBase {
+	[FIELD.store]: STORE.icon;
+	[FIELD.type]: STORE.react | STORE.class | STORE.event | STORE.provider | STORE.bonus;
+	[FIELD.key]: keyof typeof REACT | CLASS | EVENT | Auth.PROVIDER | BONUS;
+	image: string;
 	sort: number;
 }
 
 //	/client/schedule
 export interface IProvider extends IClientBase {
 	[FIELD.store]: STORE.provider;
-	[FIELD.type]: TProvider;
+	[FIELD.type]: Auth.METHOD;
 	[FIELD.key]: string;
-	sort: number;										// list-order to display to User
+	sort?: number;									// list-order to display to User
 	prefix?: string;
 	scope?: TString;								// if array, joinScope determines how to encode as a string
+	custom?: TString;								// custom OAuth parameters
 	joinScope?: string;							// what character separates the scope parameters, default ','
 	params?: object;								// custom parameters
 
@@ -291,29 +298,30 @@ export interface IProvider extends IClientBase {
 //	/member/message
 export interface IMessage extends IUserBase {
 	[FIELD.store]: STORE.message;
-	[FIELD.type]: 'diary' | 'alert';
-	[FIELD.note]: TString;
+	[FIELD.type]: MESSAGE;
 }
 
 //	/member/profile
 export interface IProfile extends IUserBase {
 	[FIELD.store]: STORE.profile;
-	[FIELD.type]: TProfile;
+	[FIELD.type]: PROFILE;
 }
 export interface IProfilePlan extends IProfile {
-	[FIELD.type]: TYPE.plan;
-	plan: TPlan;
+	[FIELD.type]: PROFILE.plan;
+	[PROFILE.plan]: PLAN;
+	bump?: PLAN;												// auto-upgrade Plan on next topUp Payment
 }
 export interface IProfileClaim extends IProfile {
-	[FIELD.type]: TYPE.claim;
-	claims: ICustomClaims;
+	[FIELD.type]: PROFILE.claim;
+	[PROFILE.claim]: ICustomClaims;
 }
 export interface IProfileInfo extends IProfile {
-	[FIELD.type]: TYPE.info;
-	[TYPE.info]: IMemberInfo;
+	[FIELD.type]: PROFILE.info;
+	[PROFILE.info]: IMemberInfo;
 }
 export interface IProfilePref extends IProfile {
-	[FIELD.type]: TYPE.pref;
+	[FIELD.type]: PROFILE.pref;
+	[PROFILE.pref]: TString;
 }
 export type TProfileInfo = IProfileInfo | IProfileInfo[];
 
@@ -332,8 +340,8 @@ export interface IMemberInfo {				// Conformed Info across Providers
 //	/member/payment
 export interface IPayment extends IUserBase {
 	[FIELD.store]: STORE.payment;
-	[FIELD.type]: TPayment;
-	link?: {
+	[FIELD.type]: PAYMENT;
+	link?: {													// used during AttendService to determine Payment hierarchy
 		child?: string;									// the child Payment
 		parent?: string;								// the parent Payment
 		index?: number;									// the Payment position
@@ -369,9 +377,9 @@ export interface IAttend extends IUserBase {
 	},
 	timetable: {
 		[FIELD.id]: string;							// the /client/schedule or /client/event _id
-		[FIELD.store]: string;					// 'schedule' or 'calendar'
-		[FIELD.type]: TSchedule;
-		[FIELD.key]: TClass;
+		[FIELD.store]: STORE.schedule | STORE.calendar;
+		[FIELD.type]: SCHEDULE;
+		[FIELD.key]: CLASS;
 	}
 	bonus?: {
 		[FIELD.id]: string;
@@ -386,21 +394,27 @@ export interface IAttend extends IUserBase {
 	}
 }
 
-//	/member/feedback
-export interface IFeedBack extends IUserBase {
-	[FIELD.store]: STORE.feedback;
-	[FIELD.type]: STORE.attend;				// currently only 'Attend' feedback
-	[FIELD.key]: string;							// key to document the Member is referencing
-	date: number;											// yyyymmdd of Feedback
-	react?: REACT;
-	comment?: TString;
+//	/forum/comment
+export interface IComment extends IForumBase {
+	[FIELD.store]: STORE.comment;
+	comment: TString;
+	response?: {
+		uid: string;
+		stamp: number;
+		comment: TString;
+		seen?: boolean;
+	}[];
+}
+//	/forum/react
+export interface IReact extends IForumBase {
+	[FIELD.store]: STORE.react;
+	react: REACT;
 }
 
 //	/admin/register
-type TRole = 'admin' | 'member' | 'guest';
 export interface ICustomClaims {		// a special sub-set of fields from the User Token
 	alias?: string;
-	roles?: TRole[];
+	roles?: Auth.ROLE[];
 	allow?: string[];
 	deny?: string[];
 }
@@ -410,7 +424,7 @@ export type TUser = UserInfo & {
 }
 type TProviderInfo = {
 	id: string;
-	provider: 'fb' | 'g+' | 'gh' | 'tw' | 'li';
+	provider: Auth.PROVIDER;
 	firstName: string;
 	lastName: string;
 	userName: string;
@@ -438,14 +452,14 @@ export interface IRegister extends IUserBase {
 //	/member/status
 export interface IStatus extends IUserBase {
 	[FIELD.store]: STORE.status;
-	[FIELD.type]: TStatus;
+	[FIELD.type]: STATUS;
 }
 export interface IStatusAccount extends IStatus {
-	[FIELD.type]: 'account';
+	[FIELD.type]: STATUS.account;
 	summary: ISummary;
 }
 export interface IStatusConnect extends IStatus {
-	[FIELD.type]: 'connect';
+	[FIELD.type]: STATUS.connect;
 	state: CONNECT;
 	device?: string | null;
 }
