@@ -12,9 +12,9 @@ export interface IWhere {
 export type TWhere = IWhere | IWhere[];
 
 /** Query.orderBy */
-interface IOrderBy {
+export interface IOrderBy {
 	fieldPath: string | FieldPath;
-	directionStr?: 'desc' | 'asc';
+	directionStr?: firebase.firestore.OrderByDirection;
 }
 /** Collection.Query */
 export interface IQuery {
