@@ -1,8 +1,8 @@
-import { State, Action, StateContext, NgxsOnInit, Store } from '@ngxs/store';
+import { State, Action, StateContext, NgxsOnInit } from '@ngxs/store';
 import { SetAdmin, DelAdmin, TruncAdmin, filterState } from '@dbase/state/state.action';
-import { TStateSlice, SLICE } from '@dbase/state/state.define';
+import { TStateSlice } from '@dbase/state/state.define';
 
-import { FIELD } from '@dbase/data/data.define';
+import { FIELD, COLLECTION } from '@dbase/data/data.define';
 import { IStoreMeta } from '@dbase/data/data.schema';
 
 import { asArray } from '@lib/array.library';
@@ -10,10 +10,10 @@ import { cloneObj } from '@lib/object.library';
 import { dbg } from '@lib/logger.library';
 
 /**
- * AdminState is for items only Users with roles['admin'] in their customClaims can see
+ * AdminState is for items that only Admin Users (with roles['admin'] in their customClaims) can see
  */
 @State<TStateSlice<IStoreMeta>>({
-	name: SLICE.admin,
+	name: COLLECTION.admin,
 	defaults: {}
 })
 export class AdminState implements NgxsOnInit {
