@@ -7,12 +7,10 @@ export class SetClient {									// Add a Client object into the Store
 	static type = '[Sync Service] Set Client';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class DelClient {									// Remove a Client object from the Store
 	static type = '[Sync Service] Delete Client';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class TruncClient {								// Truncate a Client object from the Store
 	static type = '[Sync Service] Truncate Client';
 	constructor(public debug: boolean = true) { }
@@ -22,12 +20,10 @@ export class SetDevice {										// Add a Client object-clone into the local St
 	static type = '[Sync Service] Set Device';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class DelDevice {
 	static type = '[Sync Service] Delete Device';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class TruncDevice {
 	static type = '[Sync Service] Truncate Device';
 	constructor(public debug: boolean = true) { }
@@ -37,12 +33,10 @@ export class SetMember {									// Add a Member object into the Store
 	static type = '[Sync Service] Set Member';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class DelMember {									// Remove a Member object from the Store
 	static type = '[Sync Service] Delete Member';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class TruncMember {								// Truncate a Member object from the Store
 	static type = '[Sync Service] Truncate Member';
 	constructor(public debug: boolean = true) { }
@@ -52,17 +46,14 @@ export class SetAttend {									// Add Attend object into the Store
 	static type = '[Sync Service] Set Attend';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class DelAttend {									// Remove Attend object from the Store
 	static type = '[Sync Service] Delete Attend';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class TruncAttend {								// Truncate Attend object from the Store
 	static type = '[Sync Service] Truncate Attend';
 	constructor(public debug: boolean = true) { }
 }
-
 export class SyncAttend {									// new Attend written
 	static type = 'Sync Service] Sync Attend';
 	constructor(public payload: TStoreMeta) { }
@@ -72,16 +63,27 @@ export class SetAdmin {										// Add Admin object into the Store
 	static type = '[Sync Service] Set Admin';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class DelAdmin {										// Remove Admin object from the Store
 	static type = '[Sync Service] Delete Admin';
 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
 }
-
 export class TruncAdmin {									// Truncate Admin object from the Store
 	static type = '[Sync Service] Truncate Admin';
 	constructor(public debug: boolean = true) { }
 }
+
+// export class SetForum {
+// 	static type = '[Sync Service] Set Forum';
+// 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
+// }
+// export class DelForum {
+// 	static type = '[Sync Service] Delete Forum';
+// 	constructor(public payload: TStoreMeta, public debug: boolean = true) { }
+// }
+// export class TruncForum {
+// 	static type = '[Sync Service] Truncate Forum';
+// 	constructor(public debug: boolean = true) { }
+// }
 
 /** helper function to return all except nominated document */
 export const filterState = (state: TStateSlice<IStoreMeta>, payload: IStoreMeta, segment = FIELD.store) => {
