@@ -1,7 +1,14 @@
 import { Subscription } from 'rxjs';
-import { IPromise } from '@lib/utility.library';
+
 import { COLLECTION } from '@dbase/data/data.define';
-import { TruncAttend, TruncClient, TruncMember, DelMember, DelAttend, DelClient, SetClient, SetMember, SetAttend, SetDevice, DelDevice, TruncDevice } from '@dbase/state/state.action';
+
+import { SetDevice, DelDevice, TruncDevice } from '@dbase/state/state.action';
+import { SetClient, DelClient, TruncClient } from '@dbase/state/state.action';
+import { SetMember, DelMember, TruncMember } from '@dbase/state/state.action';
+import { SetAttend, DelAttend, TruncAttend } from '@dbase/state/state.action';
+import { SetAdmin, DelAdmin, TruncAdmin } from '@dbase/state/state.action';
+
+import { IPromise } from '@lib/utility.library';
 
 export interface IListen {
 	collection: COLLECTION;								// Firestore collection
@@ -10,9 +17,9 @@ export interface IListen {
 	ready: IPromise<boolean>;							// indicate when snap0 is received
 	subscribe: Subscription;							// used for turning off Subscription
 	method: {
-		setStore: typeof SetClient | typeof SetMember | typeof SetAttend | typeof SetDevice;
-		delStore: typeof DelClient | typeof DelMember | typeof DelAttend | typeof DelDevice;
-		truncStore: typeof TruncClient | typeof TruncMember | typeof TruncAttend | typeof TruncDevice;
+		setStore: typeof SetClient | typeof SetMember | typeof SetAttend | typeof SetDevice | typeof SetAdmin;
+		delStore: typeof DelClient | typeof DelMember | typeof DelAttend | typeof DelDevice | typeof DelAdmin;
+		truncStore: typeof TruncClient | typeof TruncMember | typeof TruncAttend | typeof TruncDevice | typeof TruncAdmin;
 	}
 };
 
