@@ -2,20 +2,24 @@ import { ITimetableState } from '@dbase/state/state.define';
 import { FIELD, BONUS } from '@dbase/data/data.define';
 import { IGift, TBonus, IBonus } from '@dbase/data/data.schema';
 
-import { TDate, getDate, Instant } from '@lib/date.library';
-import { asArray } from '@lib/array.library';
+import { TDate, getDate, Instant } from '@lib/instant.library';
 import { TString, isUndefined } from '@lib/type.library';
+import { asArray } from '@lib/array.library';
 
 /**
- * Determine if provided event is entitled to a Gift or a Bonus Scheme
+ * Determine if provided event is entitled to a Gift or a Bonus  
+ * source:	info about the effective Plans, Bonus, etc. and current Member
+ * event:	  a Class name
+ * date:		effective date, else today
+ * elect:		Bonus override (e.g. Member may elect to *not* use one of their Gifts)
  */
 export const calcBonus = (source: ITimetableState, event: string, date?: TDate, elect?: BONUS) => {
 	const now = getDate(date);
-	const upd: IGift[] = [];																// an array of updates to Gifts
+	const upd: IGift[] = [];																// an array of updates for caller to apply on Gifts
 	let bonus = {} as TBonus;																// calculated Bonus entitlement
 
 	const gifts = source.member.gift;												// the active Gifts for this Member
-	const plans = source.client.plan || [];
+	const plan = (source.client.plan || [])[0];							// Member's current plan
 	const { attendGift = [], attendWeek = [], attendMonth = [], attendToday = [] } = source.attend;
 	const scheme = (source.client.bonus || [])
 		.reduce((acc, row) => {
@@ -25,9 +29,14 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 
 	switch (true) {
 		/**
-		 * Admin adds 'Gift' records to a Member's account which will detail
-		 * the start-date and limit of free classes (and an optional expiry-date for the Gift).  
+		 * Admin adds 'Gift' records to a Member's account which will detail the start-date
+		 * and limit of free classes (and an optional expiry-date for the Gift).  
 		 * These gifts will be used in preference to any other Bonus scheme.
+		 * 
+		 * Even though the Member may have an active Gift, we check if any of them are useable.  
+		 * If not, we pass back an array of auto-expire updates for the caller to apply.
+		 * 
+		 * If ok to use, we pass back an array of count-tracking updates for the caller to apply.
 		 */
 		case gifts.length > 0																	// if Member has some open Gifts
 			&& (isUndefined(elect) || elect === BONUS.gift):		//   and has not elected to *skip* this Bonus
@@ -61,8 +70,8 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 			if (curr > -1)																			// processed a Gift
 				break;																						// exit switch-case
 
-		case !plans[0].bonus:																	// Member's Plan does not allow Bonus
-			break;																							//	so no further checking
+		case !plan.bonus:																			// Member's Plan does not allow Bonus
+			break;																							//	so no further checking needed
 
 		/**
 		 * The Week scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.week.level).  
@@ -88,8 +97,8 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 			break;
 
 		/**
-		 * The Sunday scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.sunday.level),
-		 * and did not qualify for the Week scheme previously (e.g. already claimed yesterday?).    
+		 * The Sunday scheme qualifies as a Bonus if the Member attends the required number of non-bonus classes in a week (scheme.sunday.level).  
+		 * Note: the Week scheme takes precendence (if qualifies on Sunday, and not claimed on Saturday)
 		 * The class must be in the free list (scheme.week.free) to qualify for this bonus
 		 */
 		case scheme.sunday
@@ -97,9 +106,9 @@ export const calcBonus = (source: ITimetableState, event: string, date?: TDate, 
 				.filter(row => isUndefined(row.bonus) || row.bonus[FIELD.type] === BONUS.gift)
 				.distinct(row => row.track[FIELD.date])						// de-dup by day-of-week
 				.length >= scheme.sunday.level										// required number of Attends this week
-			&& now.dow === Instant.DAY.Sun											// today is 'Sunday'
+			&& now.dow === Instant.WEEKDAY.Sun									// today is 'Sunday'
 			&& elect == BONUS.sunday														// Member elected to take Bonus
-			&& (asArray(scheme.sunday.free as TString)).includes(event):
+			&& asArray(scheme.sunday.free as TString).includes(event):
 
 			bonus = {
 				[FIELD.id]: scheme.sunday[FIELD.id],
